Move resize to dataobjectfilters
